migrated "bot" subcommand to "tgcli.cli.bot" package
