+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "python.linting.enabled": false,
+    "python.linting.pep8Enabled": true,
+    "python.linting.pylintEnabled": false
+}
